+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Configure grafana-agent config to grab Kubernetes Events and send them to Loki.
+- Create grafana-agent extra secret to store logging write credentials.
 
 ## [0.1.3] - 2023-10-18