+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.4] - 2023-10-31
+
 ### Changed
 
 - Push to CAPA app collection.
@@ -106,7 +108,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add '--vintage' toggle
 - Add controller for Workload Management Cluster using cluster.x-k8s.io/v1beta1```
 
-[Unreleased]: https://github.com/giantswarm/logging-operator/compare/v0.1.3...HEAD
+[Unreleased]: https://github.com/giantswarm/logging-operator/compare/v0.1.4...HEAD
+[0.1.4]: https://github.com/giantswarm/logging-operator/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/giantswarm/logging-operator/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/giantswarm/logging-operator/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/giantswarm/logging-operator/compare/v0.1.0...v0.1.1